backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.1.0
bleach==6.1.0
sse-starlette==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.34.0
httpx==0.26.0

# Database
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Thumbnails
playwright==1.44.0
Pillow==10.2.0
diff-match-patch==20230430
